feat: filter sections by course this fixes #87
